@@ -1,9 +1,9 @@
 package com.mcamelo.simiosHuman.resources.exceptions;
 
-import java.time.Instant;
-
-import javax.servlet.http.HttpServletRequest;
-
+import com.mcamelo.simiosHuman.services.exceptions.DatabaseException;
+import com.mcamelo.simiosHuman.services.exceptions.InvalideMatrixException;
+import com.mcamelo.simiosHuman.services.exceptions.ResourceNotFoundException;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.FieldError;
@@ -11,9 +11,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
-import com.mcamelo.simiosHuman.services.exceptions.DatabaseException;
-import com.mcamelo.simiosHuman.services.exceptions.InvalideMatrixException;
-import com.mcamelo.simiosHuman.services.exceptions.ResourceNotFoundException;
+import java.time.Instant;
 
 @ControllerAdvice
 public class ResourceExceptionHandler {
